chore(#610): move images to src/assets and switch to webp/svg
